See description.
Originally this was made for someone who already had a crates plugin, but it was in russian and he wanted different features, so he asked me to modify it. Instead I ended up remaking it completely, and ended up with this. That is why some of the things in it are unfinished and/or specifically made for a single server. Most of it is my code, but there are a couple of blocks that aren't. It will probably need a lot of modifying, bug fixing and updating before it is ready to be used.